requirements.txt
cryptography~=44.0.0
os
pip~=24.3.1
torch~=2.5.1
fastapi~=0.115.6
pydantic~=2.10.3
uvicorn~=0.32.1
annotated-types~=0.7.0
SQLAlchemy~=2.0.36
PyMySQL~=1.1.1
tokenizers~=0.21.0
bcrypt~=4.2.1
transformers~=4.47.0
dotenv
datetime
pyjwt
bulma
react
ReactDOM